Praticas/Lab05/Lab05_2/pyproject.toml
[tool.poetry]
name = "lab05-2"
version = "0.1.0"
description = ""
authors = ["Pedro Pinto <pmap@ua.pt>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
kafka-python = "^2.0.2"
six = "^1.16.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
